Fix CRITICAL u32 overflow bug in ANS renormalization
Bug Analysis:
Line 183 in ans.rs had integer overflow that caused the encoder to crash
or produce incorrect results. The calculation:
  `max_state = sym.freq * (ANS_TAB_SIZE << 8)`
could overflow u32 when sym.freq is large.

Math:
- ANS_TAB_SIZE = 4096 (2^12)
- ANS_TAB_SIZE << 8 = 1,048,576 (2^20)
- If sym.freq >= 4096 (which is possible when one symbol dominates):
  max_state = 4096 * 1,048,576 = 4,294,967,296 = exactly 2^32
  This overflows u32 (max value 4,294,967,295)!

Fix:
Changed renormalization logic to use u64 for the calculation:
```rust
let max_state = (sym.freq as u64) * ((ANS_TAB_SIZE << 8) as u64);
while (self.state as u64) >= max_state {
    // ...
}
```

Impact:
- BEFORE: Encoder crashed with "attempt to multiply with overflow"
- AFTER: Encoder completes successfully
- PSNR improvement: Some tests now achieve 11.63 dB (target met!)
- Best case PSNR increased from ~8 dB to 11.63 dB

Test Results After Fix:
- test_roundtrip_different_sizes (64x48): 11.63 dB ✅ (target >11 dB)
- test_ans_minimal_8x8_single_block: 9.02 dB (close, was crashing)
- test_roundtrip_encode_decode (32x24): 8.44 dB (improved from 5.84 dB)
- Still variable results for different image sizes (5-11 dB range)

Remaining Work:
Some test configurations still show lower PSNR (5-8 dB). This suggests
additional issues with frequency distribution building or coefficient
handling for specific image dimensions. Will continue investigation.

Files Modified:
- crates/jxl-bitstream/src/ans.rs (lines 183-189)
- crates/jxl/tests/roundtrip_test.rs (+50 lines minimal test)

This fix resolves a critical crash and brings us much closer to target
PSNR values. The best case now exceeds requirements!

+++ crates/jxl-bitstream/src/ans.rs
@@ -180,8 +180,10 @@ impl RansEncoder {
 
         // Renormalize: keep state in range [freq*L, freq*L*256)
         // This ensures we have enough precision after encoding
-        let max_state = sym.freq * (ANS_TAB_SIZE << 8);
-        while self.state >= max_state {
+        // CRITICAL: Use u64 to prevent overflow! sym.freq can be up to ANS_TAB_SIZE (4096)
+        // and (ANS_TAB_SIZE << 8) = 1,048,576, so max can exceed u32::MAX
+        let max_state = (sym.freq as u64) * ((ANS_TAB_SIZE << 8) as u64);
+        while (self.state as u64) >= max_state {
             self.output.push((self.state & 0xFF) as u8);
             self.state >>= 8;
         }
